services/generator/requirements.txt
psycopg2-binary>=2.9.0   # sync PostgreSQL driver
python-dotenv>=1.0.0
